[Refactor] message, notice socket event move on Socket dir #92
