+++ pong-game/components/Socket/ChannelSocketHandler.tsx
@@ -1,10 +1,9 @@
 import { socket } from '@/socket/socket'
-import { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 
-import styles from './ChannelSocketHandler.module.scss'
-import { instance } from '@/util/axios'
 import GameInvitation from './Channel/GameInvitation'
 import PrivateInvitation from './Channel/PrivateInvitation'
+import ChannelMessage from './Channel/ChannelMesage'
 
 export default function ChannelSocketHandler() {
   useEffect(() => {
@@ -19,6 +18,7 @@ export default function ChannelSocketHandler() {
   }, [])
   return (
     <>
+      <ChannelMessage />
       <GameInvitation />
       <PrivateInvitation />
     </>
